+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.2] - 2026-02-19
+
+- Add `AsepriteFile::render_frame` method which supports overriding thei
+  layer visibility via a `LayerSelection`.
+- Deprecate `AsepriteFile::combined_frame_image` in favor of
+  `AsepriteFile::render_frame`
+
+## [0.4.1] - 2026-02-19
+
+- Fix size bug when parsing properties maps
+
 ## [0.4.0] - 2025-07-21
 
 ## Added
@@ -98,7 +109,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/bikeshedder/aseprite-loader/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/bikeshedder/aseprite-loader/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/bikeshedder/aseprite-loader/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/bikeshedder/aseprite-loader/compare/v0.3.4...v0.4.0
 [0.3.4]: https://github.com/bikeshedder/aseprite-loader/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/bikeshedder/aseprite-loader/compare/v0.3.2...v0.3.3